[BE] refactor(#87): 피드백 반영
* 이메일 검증 메서드 유틸 클래스로 분리
* Long 타입 비교 Objects.equals에서 != 비교 연산자로 변경
